refactored meeting functions out into a separate file
